requirements.txt
streamlit==0.62.1
spacy==2.3.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz
textblob==0.15.3
matplotlib
wordcloud ==1.7.0
bs4==0.0.1
gensim==3.8.3
sumy==0.8.1
googletrans==3.0.0
nltk==3.5
Pillow>=6.2.2